swagger added creation orders and payments works

+++ src/main/kotlin/se/nt1dc/deliveryservice/dto/LocationDto.kt
@@ -1,6 +1,6 @@
 package se.nt1dc.deliveryservice.dto
 
 data class LocationDto(
-    var latitude: Long,
-    var longitude: Long,
+    var latitude: Double,
+    var longitude: Double,
 )
